[ALPS02361141] WMT:out-of-bounds access
[Detail]
    Pointer &pData->data points to an object whose effective type
is unsigned int but is dereferenced as a wider unsigned long. This may
lead to memeory corruption.

[Solution] change the type of Pointer &pData->data unsigned int to
unsigned long.

MTK-Commit-Id: fc0c81d8f1e800f84464d1fa4b68194e1ceb5b32

CR-Id: ALPS02361141
Feature: WMT

+++ drivers/misc/mediatek/connectivity/common/combo/linux/include/osal.h
@@ -227,7 +227,7 @@ typedef struct _OSAL_WAKE_LOCK_ {
 } OSAL_WAKE_LOCK, *P_OSAL_WAKE_LOCK;
 #if 1
 typedef struct _OSAL_BIT_OP_VAR_ {
-	UINT32 data;
+	unsigned long data;
 	OSAL_UNSLEEPABLE_LOCK opLock;
 } OSAL_BIT_OP_VAR, *P_OSAL_BIT_OP_VAR;
 #else
